+++ src/RutokenPkcs11Interop/HighLevelAPI41/Factories/RutokenInitParamFactory.cs
@@ -4,6 +4,8 @@ using Net.RutokenPkcs11Interop.Common;
 using Net.RutokenPkcs11Interop.HighLevelAPI.Factories;
 using Net.RutokenPkcs11Interop.HighLevelAPI;
 
+// Note: Code in this file is maintained manually
+
 namespace Net.RutokenPkcs11Interop.HighLevelAPI41.Factories
 {
     /// <summary>
@@ -12,16 +14,16 @@ namespace Net.RutokenPkcs11Interop.HighLevelAPI41.Factories
     public class RutokenInitParamFactory : IRutokenInitParamFactory
     {
         public IRutokenInitParam Create(string newAdminPin, string newUserPin, string tokenLabel,
-            IList<RutokenFlag> changeUserPINPolicy, uint minAdminPinLen, uint minUserPinLen,
-            uint maxAdminRetryCount, uint maxUserRetryCount, uint smMode)
+            IList<RutokenFlag> changeUserPINPolicy, ulong minAdminPinLen, ulong minUserPinLen,
+            ulong maxAdminRetryCount, ulong maxUserRetryCount, ulong smMode)
         {
             return new RutokenInitParam(newAdminPin, newUserPin, tokenLabel, changeUserPINPolicy, minAdminPinLen, minUserPinLen, 
                 maxAdminRetryCount, maxUserRetryCount, smMode);
         }
 
         public IRutokenInitParam Create(string newAdminPin, string newUserPin, string tokenLabel,
-            IList<RutokenFlag> changeUserPINPolicy, uint minAdminPinLen, uint minUserPinLen,
-            uint maxAdminRetryCount, uint maxUserRetryCount, uint smMode, bool useRepairMode)
+            IList<RutokenFlag> changeUserPINPolicy, ulong minAdminPinLen, ulong minUserPinLen,
+            ulong maxAdminRetryCount, ulong maxUserRetryCount, ulong smMode, bool useRepairMode)
         {
             return new RutokenInitParam(newAdminPin, newUserPin, tokenLabel, changeUserPINPolicy, minAdminPinLen, minUserPinLen,
             maxAdminRetryCount, maxUserRetryCount, smMode, useRepairMode);
